requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.27.0
tenacity>=9.0.0

# LLM / agent
langgraph>=0.2.50
langchain-core>=0.3.20
openai>=1.55.0

# Retrieval
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
numpy>=1.26.0
rank-bm25>=0.2.2

# Scraping / parsing
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Frontend
streamlit>=1.40.0

# Eval / utils
pandas>=2.2.0
orjson>=3.10.0
matplotlib>=3.9.0

# Optional live scrape (Cloudflare bypass): pip install playwright && playwright install chromium
# playwright>=1.49.0
